backend/requirements.txt
Flask
Werkzeug
markupsafe
flask-sqlalchemy
flask-marshmallow
Flask-Cors
python-dotenv
SQLAlchemy